# packages/api-schema/openapi.yaml
openapi: 3.1.0
info:
  title: WISE Workbench API
  description: 'WISE Workbench backend: projects, datasets, mappings, case tables, norm versions, runs, backlogs, slice detail, traces, diagnostics, signal distributions, process flow and jobs. Errors are RFC 9457 problem details.'
  version: 0.1.0
paths:
...
  /projects/{projectId}/runs:
    get:
      tags:
      - runs
      summary: List Runs
      operationId: listRuns
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Run'
                title: Response Listruns
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - runs
      summary: Create Run
      description: Creates a run and queues its job; identical parameters (or the same Idempotency-Key) return the existing run.
      operationId: createRun
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: force
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Force
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunCreate'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/runs/{runId}:
    get:
      tags:
      - runs
      summary: Get Run
      operationId: getRun
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/runs/{runId}/cancel:
    post:
      tags:
      - runs
      summary: Cancel Run
      operationId: cancelRun
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Run'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/runs/{runId}/summary:
    get:
      tags:
      - runs
      summary: Get Run Summary
      operationId: getRunSummary
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/runs/{runId}/backlog:
    get:
      tags:
      - runs
      summary: Get Backlog
      operationId: getBacklog
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      - name: slicing
        in: query
        required: true
        schema:
          type: string
          description: slicing id or comma-separated case attributes
          title: Slicing
        description: slicing id or comma-separated case attributes
      - name: view
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: View
      - name: gamma
        in: query
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          title: Gamma
      - name: minCases
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 20
          title: Mincases
      - name: sort
        in: query
        required: false
        schema:
          type: string
          default: -PI
          title: Sort
      - name: hotspotType
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - severity
            - mechanism
            - reservoir
            type: string
          - type: 'null'
          description: method name of the kind
          title: Hotspottype
        description: method name of the kind
      - name: kind
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - acute
            - systematic
            - widespread
            type: string
          - type: 'null'
          description: kind of problem
          title: Kind
        description: kind of problem
      - name: layer
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: most-missed expectation area (layer id)
          title: Layer
        description: most-missed expectation area (layer id)
      - name: stability
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - stable
            - fragile
            - insufficient_support
            - unknown
            type: string
          - type: 'null'
          description: confidence badge
          title: Stability
        description: confidence badge
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - name: bands
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'JSON list of band specs for numeric attributes: [{"attribute": "exposure", "method": "quantile", "q": 4}]'
          title: Bands
        description: 'JSON list of band specs for numeric attributes: [{"attribute": "exposure", "method": "quantile", "q": 4}]'
      - name: drillFrom
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'drill into one group of this slicing (id or attributes): a finer backlog scoped to it'
          title: Drillfrom
        description: 'drill into one group of this slicing (id or attributes): a finer backlog scoped to it'
      - name: drillKey
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: the group's key (JSON array) in drillFrom
          title: Drillkey
        description: the group's key (JSON array) in drillFrom
      - name: filter
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'URL-safe JSON filter: {"and": [{"kind": "attribute", "field": "…", "in": […]}, …]}; never changes applicability'
          title: Filter
        description: 'URL-safe JSON filter: {"and": [{"kind": "attribute", "field": "…", "in": […]}, …]}; never changes applicability'
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 50
          title: Pagesize
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacklogPage'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/runs/{runId}/slicings/preview:
    get:
      tags:
      - runs
      summary: Preview Slicing
      description: 'The slice designer''s preview: how many groups a slicing (two or three attributes, banded numbers) makes and how big they are.'
      operationId: previewSlicing
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      - name: slicing
        in: query
        required: true
        schema:
          type: string
          description: slicing id or comma-separated case attributes
          title: Slicing
        description: slicing id or comma-separated case attributes
      - name: bands
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'JSON list of band specs for numeric attributes: [{"attribute": "exposure", "method": "quantile", "q": 4}]'
          title: Bands
        description: 'JSON list of band specs for numeric attributes: [{"attribute": "exposure", "method": "quantile", "q": 4}]'
      - name: minCases
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 20
          title: Mincases
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlicingPreview'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/runs/{runId}/filters/preview:
    get:
      tags:
      - runs
      summary: Preview Filter
      description: Cases in and out of a filter, what each clause removes on its own, and cases in scope per expectation.
      operationId: previewFilter
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      - name: filter
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'URL-safe JSON filter: {"and": [{"kind": "attribute", "field": "…", "in": […]}, …]}; never changes applicability'
          title: Filter
        description: 'URL-safe JSON filter: {"and": [{"kind": "attribute", "field": "…", "in": […]}, …]}; never changes applicability'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilterPreview'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/runs/{runId}/analytics:
    get:
      tags:
      - runs
      summary: Get Analytics
      description: Whether the run's analytics (stability, kinds, comparisons, caveats, readiness gate) are computed, with the manifest of cached records.
      operationId: getAnalytics
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsStatus'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - runs
      summary: Request Analytics
      description: Queue the analytics job for a finished run (it runs after every scoring job by default).
      operationId: requestAnalytics
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/runs/{runId}/compare-flow-types:
    get:
      tags:
      - runs
      summary: Compare Flow Types
      description: 'Per flow type side by side: cases, score and shortfall per view, the most-missed expectation and the top groups (a run without scope).'
      operationId: compareFlowTypes
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      - name: attribute
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'attribute holding the flow type (default: the mapping''s)'
          title: Attribute
        description: 'attribute holding the flow type (default: the mapping''s)'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowTypeComparison'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/runs/{runId}/slices/{sliceKey}:
    get:
      tags:
      - runs
      summary: Get Slice
      operationId: getSlice
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      - name: sliceKey
        in: path
        required: true
        schema:
          type: string
          title: Slicekey
      - name: slicing
        in: query
        required: true
        schema:
          type: string
          title: Slicing
      - name: view
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: View
      - name: drilldown
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: case attribute for the penalty-mass Pareto
          title: Drilldown
        description: case attribute for the penalty-mass Pareto
      - name: bands
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'JSON list of band specs for numeric attributes: [{"attribute": "exposure", "method": "quantile", "q": 4}]'
          title: Bands
        description: 'JSON list of band specs for numeric attributes: [{"attribute": "exposure", "method": "quantile", "q": 4}]'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SliceDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/runs/{runId}/cases/{caseId}/trace:
    get:
      tags:
      - runs
      summary: Get Trace
      operationId: getTrace
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      - name: caseId
        in: path
        required: true
        schema:
          type: string
          title: Caseid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trace'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/runs/{runId}/diagnostics:
    get:
      tags:
      - runs
      summary: Get Diagnostics
      operationId: getDiagnostics
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      - name: slicing
        in: query
        required: true
        schema:
          type: string
          title: Slicing
      - name: view
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: View
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Table'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /projects/{projectId}/runs/{runId}/signals/{constraintId}:
    get:
      tags:
      - runs
      summary: Get Signal Distribution
      operationId: getSignalDistribution
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
          title: Projectid
      - name: runId
        in: path
        required: true
        schema:
          type: string
          title: Runid
      - name: constraintId
        in: path
        required: true
        schema:
          type: string
          title: Constraintid
      - name: slicing
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Slicing
      - name: sliceKey
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Slicekey
      - name: filter
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'URL-safe JSON filter: {"and": [{"kind": "attribute", "field": "…", "in": […]}, …]}; never changes applicability'
          title: Filter
        description: 'URL-safe JSON filter: {"and": [{"kind": "attribute", "field": "…", "in": […]}, …]}; never changes applicability'
      - name: scale
        in: query
        required: false
        schema:
          enum:
          - linear
          - log
          type: string
          description: bin scale of the histogram
          default: linear
          title: Scale
        description: bin scale of the histogram
      - name: bands
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'JSON list of band specs for numeric attributes: [{"attribute": "exposure", "method": "quantile", "q": 4}]'
          title: Bands
        description: 'JSON list of band specs for numeric attributes: [{"attribute": "exposure", "method": "quantile", "q": 4}]'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Distribution'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Run:
      properties:
        caseTableId:
          type: string
          title: Casetableid
        normVersionId:
          type: string
          title: Normversionid
        views:
          items:
            type: string
          type: array
          title: Views
        slicings:
          items:
            $ref: '#/components/schemas/SlicingSpec'
          type: array
          title: Slicings
        gamma:
          type: number
          title: Gamma
          default: 50
        minCases:
          type: integer
          title: Mincases
          default: 20
        baselineRunId:
          anyOf:
          - type: string
          - type: 'null'
          title: Baselinerunid
        note:
          anyOf:
          - type: string
          - type: 'null'
          title: Note
        scope:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Scope
          description: the run's sub-log scope as recorded
        id:
          type: string
          title: Id
        status:
          type: string
          enum:
          - queued
          - running
          - done
          - failed
          - cancelled
          title: Status
        jobId:
          anyOf:
          - type: string
          - type: 'null'
          title: Jobid
        manifest:
          anyOf:
          - $ref: '#/components/schemas/RunManifest'
          - type: 'null'
        paramsHash:
          type: string
          title: Paramshash
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
        createdAt:
          type: string
          format: date-time
          title: Createdat
        links:
          additionalProperties:
            type: string
          type: object
          title: Links
      type: object
      required:
      - caseTableId
      - normVersionId
      - id
      - status
      - paramsHash
      - createdAt
      title: Run
...
servers:
- url: /api/v1